third_party/blink/web_tests/platform/generic/http/tests/misc/webtiming-cross-origin-and-back-expected.txt

If the destination and previous page have the same origin, then Web Timing should report unload timing.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS timing.connectEnd is non-zero.
PASS timing.connectStart is non-zero.
PASS timing.domComplete is non-zero.
PASS timing.domContentLoadedEventEnd is non-zero.
PASS timing.domContentLoadedEventStart is non-zero.
PASS timing.domInteractive is non-zero.
PASS timing.domLoading is non-zero.
PASS timing.domainLookupEnd is non-zero.
PASS timing.domainLookupStart is non-zero.
PASS timing.fetchStart is non-zero.
PASS timing.loadEventEnd is 0
PASS timing.loadEventStart is non-zero.
PASS timing.navigationStart is non-zero.
PASS timing.redirectEnd is 0
PASS timing.redirectStart is 0
PASS timing.requestStart is non-zero.
PASS timing.responseEnd is non-zero.
PASS timing.responseStart is non-zero.
PASS timing.secureConnectionStart is 0
PASS timing.toJSON is non-zero.
PASS timing.unloadEventEnd is 0
PASS timing.unloadEventStart is 0
PASS navigation.redirectCount is 0
PASS successfullyParsed is true

TEST COMPLETE
